fix : 에러메세지 수정
에러 catch시 [ERROR] 문자열 포함
